+++ src/main/java/madsciencemod/cards/FoldedTrinket.java
@@ -14,6 +14,7 @@ public class FoldedTrinket extends AbstractTrinket {
     private static final CardStrings cardStrings = CardCrawlGame.languagePack.getCardStrings(ID);
     public static final String NAME = cardStrings.NAME;
     public static final String DESCRIPTION = cardStrings.DESCRIPTION;
+    public static final String UPGRADE_DESCRIPTION = cardStrings.UPGRADE_DESCRIPTION;
     private static final int COST = 0;
     private static final int DRAW_AMT = 2;
     private static final int UPGRADE_AMT = 1;
@@ -40,6 +41,8 @@ public class FoldedTrinket extends AbstractTrinket {
         if (!this.upgraded) {
             upgradeName();
             upgradeMagicNumber(UPGRADE_AMT);
+            this.rawDescription = UPGRADE_DESCRIPTION;
+            this.initializeDescription();
         }
     }
 }