+++ src/main/java/madsciencemod/cards/Probe.java
@@ -22,7 +22,7 @@ public class Probe extends AbstractMadScienceCard {
     private static final int DAMAGE = 7;
     private static final int UPGRADE_DAMAGE = 3;
     private static final int CARDS = 2;
-    private static final int UPGRADE_CARDS = 2;
+    private static final int UPGRADE_CARDS = 1;
     private static final CardType TYPE = CardType.ATTACK;
     private static final CardRarity RARITY = CardRarity.UNCOMMON;
     private static final CardTarget TARGET = CardTarget.ENEMY;