+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Changed Now or Never: gives 11(14) block next turn instead of 12(16)
 * Changed Box of Scraps: draws before shuffling trinkets
 * Changed Probe: look at 2(3) cards instead of 2(4) cards
+* Changed Redesign: draw 4(5) cards instead of 4(6) cards
 * Removed Nitro Boost, it was never a collectible card to begin with
 
 ### v0.1.1